# web/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CYNA API",
    "description": "API REST pour la gestion du système CYNA",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "API via proxy"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id_utilisateur": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["client", "admin"]
          },
          "est_actif": {
            "type": "boolean"
          },
          "totp_enabled": {
            "type": "boolean"
          }
        }
      },
...
      "CarouselImage": {
        "type": "object",
        "properties": {
          "id_image": {
            "type": "integer"
          },
          "titre": {
            "type": "string"
          },
          "url_image": {
            "type": "string"
          },
          "ordre_affichage": {
            "type": "integer"
          },
          "est_actif": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Liste des utilisateurs",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Créer un utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Détails d'un utilisateur",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de l'utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Mettre à jour un utilisateur",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur mis à jour"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Supprimer un utilisateur",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur supprimé"
          }
        }
      }
    },
    "/api/users/{id}/reset-2fa": {
      "post": {
        "tags": ["Users"],
        "summary": "Réinitialiser le 2FA d'un utilisateur (Admin uniquement)",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "2FA réinitialisé avec succès"
          },
          "403": {
            "description": "Accès refusé - Admin requis"
          }
        }
      }
    },
...
    "/api/carousel-images": {
      "get": {
        "tags": ["Carousel"],
        "summary": "Liste des images du carrousel",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Liste des images",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CarouselImage"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Carousel"],
        "summary": "Créer une image du carrousel",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarouselImage"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image créée"
          }
        }
      }
    },
    "/api/carousel-images/{id}": {
      "get": {
        "tags": ["Carousel"],
        "summary": "Détails d'une image",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de l'image"
          }
        }
      },
      "put": {
        "tags": ["Carousel"],
        "summary": "Mettre à jour une image",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarouselImage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image mise à jour"
          }
        }
      },
      "delete": {
        "tags": ["Carousel"],
        "summary": "Supprimer une image",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image supprimée"
          }
        }
      }
    },
    "/api/carousel-images/reorder": {
      "post": {
        "tags": ["Carousel"],
        "summary": "Réordonner les images du carrousel",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id_image": {
                          "type": "integer"
                        },
                        "ordre_affichage": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Images réordonnées"
          }
        }
      }
    },
...
    "/api/public/products/{slug}": {
      "get": {
        "tags": ["Public"],
        "summary": "Liste publique des produits d'une catégorie",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des produits"
          }
        }
      }
    }
  }
}